refactor: TotalItemCount 로직(Header 랜더링 로직) 분리
